시트 처리 로직 리팩토링 및 헤더 정리
- `SearchInSelectedFiles()` 시트 순회 로직 개선 및 디버깅 출력 방식 변경
- `goto` 제거, `bSkipSheet` 플래그와 `continue`로 시트 스킵 처리
- 셀 순회 반복 변수와 `ProcessCell()` 매개변수를 `uint16_t`로 통일

+++ ExcelSearcher/ImGuiManager.h
@@ -54,5 +54,5 @@ private:
     // 엑셀 파일 검색 처리
     void SearchInSelectedFiles(const std::string& keyword);  // 선택된 파일들에서 검색
     std::string CopyExcelFile(const std::string& originalPath); // 엑셀 파일 복사
-    bool ProcessCell(OpenXLSX::XLWorksheet& sheet, const std::string& fileName, const std::string& sheetName, const std::string& keyword, uint64_t row, uint16_t col);
+    bool ProcessCell(OpenXLSX::XLWorksheet& sheet, const std::string& fileName, const std::string& sheetName, const std::string& keyword, uint16_t row, uint16_t col);
 };
